requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
black==23.10.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
psycopg2-binary==2.9.9
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.22
types-python-dateutil==2.8.19.14
typing_extensions==4.7.1
yarl==1.9.2
